Logging to check internals
Current findings is that when adding a new schedule, the timeEndInvalid field of validations becomes true so it deactivates the button. However, when removing the schedule, this field does not become false again

Furthermore, the timeEndInvalid field is that only one that becomes true when a new schedule is added, meaining that one you turn on 'All Day', the thing is counted as valid and you can Add On. This is even though so start date was selected.